Move common net classes in LowLevel subdirectory
