+++ src/ScaleNet.Server/Connection.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using ScaleNet.Common;
+using ScaleNet.Common.LowLevel;
 using ScaleNet.Server.LowLevel.Transport;
 
 namespace ScaleNet.Server;
