+++ src/ScaleNet.Server/LowLevel/Transport/Tcp/Implementation/SslSession.cs
@@ -3,6 +3,7 @@ using System.Net.Security;
 using System.Net.Sockets;
 using System.Text;
 using ScaleNet.Common;
+using ScaleNet.Common.LowLevel;
 
 namespace ScaleNet.Server.LowLevel.Transport.Tcp;
